pyproject.toml
[tool.poetry]
name = "slapy"
version = "0.1.0"
description = ""
authors = ["mohira <mohira@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
slack-sdk = "^3.15.2"
python-dotenv = "^0.20.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.4"
vcrpy = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
